Changed: Moved GraphPropertyWorkerTestBase and GraphPropertyWorkerTestSetupBase to core-test
These base test classes were in a unnecessary module graph-property-worker-plugin-test
this change simplifies using these test classes

CHANGELOG
Changed: Moved GraphPropertyWorkerTestBase and GraphPropertyWorkerTestSetupBase to core-test
