--- graph-property-worker/graph-property-worker-plugin-test/src/main/resources/META-INF/services/org.visallo.core.cmdline.CommandLineTool
@@ -1,1 +1,0 @@
-org.visallo.test.TestJettyServer
