--- graph-property-worker/graph-property-worker-plugin-test/src/main/resources/org/visallo/test/test.owl
@@ -1,74 +1,0 @@
-<?xml version="1.0"?>
-
-
-<!DOCTYPE rdf:RDF [
-    <!ENTITY visallo "http://visallo.org#" >
-    <!ENTITY owl "http://www.w3.org/2002/07/owl#" >
-    <!ENTITY xsd "http://www.w3.org/2001/XMLSchema#" >
-    <!ENTITY rdfs "http://www.w3.org/2000/01/rdf-schema#" >
-    <!ENTITY rdf "http://www.w3.org/1999/02/22-rdf-syntax-ns#" >
-]>
-
-
-<rdf:RDF xmlns="http://visallo.org/test#"
-     xml:base="http://visallo.org/test"
-     xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
-     xmlns:visallo="http://visallo.org#"
-     xmlns:owl="http://www.w3.org/2002/07/owl#"
-     xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
-     xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
-    <owl:Ontology rdf:about="http://visallo.org/test">
-        <owl:imports rdf:resource="http://visallo.org"/>
-    </owl:Ontology>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Object Properties
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://visallo.org/test#hasEntity -->
-
-    <owl:ObjectProperty rdf:about="http://visallo.org/test#hasEntity">
-        <rdfs:domain rdf:resource="http://visallo.org/test#artifact"/>
-        <rdfs:range rdf:resource="&owl;Thing"/>
-    </owl:ObjectProperty>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Classes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://visallo.org/test#artifact -->
-
-    <owl:Class rdf:about="http://visallo.org/test#artifact"/>
-    
-
-
-    <!-- http://visallo.org/test#document -->
-
-    <owl:Class rdf:about="http://visallo.org/test#document">
-        <rdfs:subClassOf rdf:resource="http://visallo.org/test#artifact"/>
-        <visallo:displayType>document</visallo:displayType>
-    </owl:Class>
-</rdf:RDF>
-
-
-
-<!-- Generated by the OWL API (version 3.4.2) http://owlapi.sourceforge.net -->
-
