--- graph-property-worker/graph-property-worker-plugin-test/src/main/resources/org/visallo/test/visallo.properties
@@ -1,90 +1,0 @@
-# Application-wide
-hadoop.url=file:///tmp/visallo-integration-test
-zookeeper.serverNames=localhost
-dfs.client.use.datanode.hostname=true
-
-webAppEmbedded.longRunningProcessRunner.enabled=false
-webAppEmbedded.graphPropertyWorkerRunner.enabled=false
-
-# Object-detection
-objectdetection.classifierConcepts=face
-objectdetection.classifier.face.path=/tmp/visallo-integration-test/visallo/config/opencv/haarcascade_frontalface_alt.xml
-
-termextraction.opennlp.pathPrefix=file:///tmp/visallo-integration-test/visallo/config/opennlp/
-org.visallo.geoip.GeoIpGraphPropertyWorker.pathPrefix=file:///tmp/visallo-integration-test/visallo/config/org.visallo.geoip.GeoIpGraphPropertyWorker/
-
-#hdfsLib.sourceDirectory=hdfs://visallo-dev/visallo/lib
-#hdfsLib.tempDirectory=/tmp/hdfslib
-
-org.visallo.core.model.user.UserPropertyPrivilegeRepository.defaultPrivileges=READ,COMMENT,EDIT,PUBLISH,SEARCH_SAVE_GLOBAL,ADMIN
-org.visallo.core.model.user.UserPropertyAuthorizationRepository.defaultAuthorizations=
-
-web.cacheServletFilter.maxAge=3600
-
-# Map component
-web.ui.map.provider=osm
-web.ui.map.provider.osm.url=https://{a-c}.tile.openstreetmap.org/${z}/${x}/${y}.png
-
-# Providers
-fs.provider=org.visallo.fs.hdfs.HdfsSession
-
-tesseract.dataPath=/usr/share/tesseract-ocr/tessdata/
-
-# Repository
-simpleOrmSession=com.v5analytics.simpleorm.InMemorySimpleOrmSession
-repository.lock=org.visallo.core.model.lock.NonLockingLockRepository
-repository.user=org.visallo.vertexium.model.user.VertexiumUserRepository
-repository.search=org.visallo.vertexium.model.search.VertexiumSearchRepository
-repository.longRunningProcess=org.visallo.vertexium.model.longRunningProcess.VertexiumLongRunningProcessRepository
-repository.workspace=org.visallo.vertexium.model.workspace.VertexiumWorkspaceRepository
-repository.graphAuthorization=org.visallo.vertexium.model.user.AccumuloGraphAuthorizationRepository
-repository.ontology=org.visallo.vertexium.model.ontology.VertexiumOntologyRepository
-repository.audit=org.visallo.core.model.audit.SimpleOrmAuditRepository
-repository.workQueue=org.visallo.model.queue.inmemory.InMemoryWorkQueueRepository
-repository.email=org.visallo.core.email.NopEmailRepository
-repository.authorization=org.visallo.core.model.user.UserPropertyAuthorizationRepository
-repository.privilege=org.visallo.core.model.user.UserPropertyPrivilegeRepository
-
-# Clavin
-clavin.disabled=false
-clavin.indexDirectory=/opt/visallo/clavin-index
-# For maxHitDepth & maxContextWindow, higher values yield better accuracy, but slower performance.
-clavin.maxHitDepth=5
-clavin.maxContextWindow=5
-# Setting useFuzzyMatching to true increases recall & decreases precision (more false positives).
-clavin.useFuzzyMatching=false
-clavin.excludeIri.0=http://visallo.org/test#zipCode
-
-# Visibility Plugin
-security.visibilityTranslator=org.visallo.core.security.DirectVisibilityTranslator
-
-# Simple ORM: Accumulo specific
-simpleOrm.accumulo.instanceName=miniInstance
-simpleOrm.accumulo.zookeeperServerNames=localhost
-simpleOrm.accumulo.username=root
-simpleOrm.accumulo.password=test
-
-# Vertexium
-graph=org.vertexium.accumulo.AccumuloGraph
-# TODO when server side element visibility is fix remove this line
-graph.useServerSideElementVisibilityRowFilter=false
-graph.accumuloInstanceName=miniInstance
-graph.username=root
-graph.password=test
-graph.tableNamePrefix=visallo_vertexium
-graph.zookeeperServers=localhost
-graph.serializer=org.vertexium.JavaVertexiumSerializer
-graph.idgenerator=org.vertexium.id.UUIDIdGenerator
-graph.hdfs.rootDir=hdfs://localhost
-graph.hdfs.dataDir=/visallo/vertexium
-graph.search=org.vertexium.elasticsearch.ElasticsearchSingleDocumentSearchIndex
-graph.search.locations=localhost
-graph.search.indexName=vertexium
-
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.image.regex=image.*
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.image.intent=image
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.audio.regex=audio.*
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.audio.intent=audio
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.video.regex=video.*
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.video.intent=video
-org.visallo.mimeTypeOntologyMapper.MimeTypeOntologyMapperGraphPropertyWorker.mapping.default.intent=document
